Factor all the report generation code out
Datatypes that are involved in report generation are in
search_results.py now, and the report generation code is in search_report.py
