+++ src/search_file.py
@@ -55,6 +55,8 @@ from coq_serapy.contexts import (TacticContext, ScrapedTactic,
 from util import (unwrap, eprint, escape_filename, escape_lemma_name,
                   mybarfmt, split_by_char_outside_matching, nostderr)
 import search_report
+from search_results import (SearchResult, SearchStatus, KilledException,
+                            TacticInteraction)
 import multi_project_report
 import util
 import tokenizer
@@ -72,81 +74,6 @@ import pygraphviz as pgv
 import torch
 
 from value_estimator import Estimator
-Tag = Callable[..., Doc.Tag]
-Text = Callable[..., None]
-Line = Callable[..., None]
-
-details_css = "details.css"
-details_javascript = "search-details.js"
-
-
-class SearchStatus(str, Enum):
-    SUCCESS = 'SUCCESS'
-    INCOMPLETE = 'INCOMPLETE'
-    SKIPPED = 'SKIPPED'
-    FAILURE = 'FAILURE'
-    CRASHED = 'CRASHED'
-
-
-class VernacBlock(NamedTuple):
-    commands: List[str]
-
-
-class TacticInteraction(NamedTuple):
-    tactic: str
-    context_before: ProofContext
-
-    @classmethod
-    def from_dict(cls, data):
-        tactic = data['tactic']
-        context_before = ProofContext.from_dict(data['context_before'])
-        return cls(tactic, context_before)
-
-    def to_dict(self):
-        return {"tactic": self.tactic,
-                "context_before": self.context_before.to_dict()}
-
-
-class ProofBlock(NamedTuple):
-    lemma_statement: str
-    module: Optional[str]
-    status: SearchStatus
-    predicted_tactics: List[TacticInteraction]
-    original_tactics: List[TacticInteraction]
-
-
-class ArgsMismatchException(Exception):
-    pass
-
-
-class SourceChangedException(Exception):
-    pass
-
-class KilledException(Exception):
-    pass
-
-
-class SearchResult(NamedTuple):
-    status: SearchStatus
-    commands: Optional[List[TacticInteraction]]
-
-    @classmethod
-    def from_dict(cls, data):
-        status = SearchStatus(data['status'])
-        if data['commands'] is None:
-            commands = None
-        else:
-            commands = list(map(TacticInteraction.from_dict,
-                                data['commands']))
-        return cls(status, commands)
-
-    def to_dict(self):
-        return {'status': self.status.name,
-                'commands': list(map(TacticInteraction.to_dict,
-                                     self.commands))}
-
-
-DocumentBlock = Union[VernacBlock, ProofBlock]
 
 unnamed_goal_number: int = 0
 
@@ -164,27 +91,6 @@ def main(arg_list: List[str]) -> None:
         util.cuda_device = f"cuda:{args.gpu}"
 
     predictor = get_predictor(parser, args)
-    base = Path(os.path.dirname(os.path.abspath(__file__)))
-
-    if not args.output_dir.exists():
-        os.makedirs(str(args.output_dir))
-    if args.splits_file:
-        with args.splits_file.open('r') as splits_f:
-            project_dicts = json.loads(splits_f.read())
-        for project_dict in project_dicts:
-            os.makedirs(args.output_dir / project_dict["project_name"], exist_ok=True)
-            for filename in [details_css, details_javascript]:
-                destpath = args.output_dir / project_dict["project_name"] / filename
-                if not destpath.exists():
-                    srcpath = base.parent / 'reports' / filename
-                    copyfile(srcpath, destpath)
-    else:
-        for filename in [details_css, details_javascript]:
-            destpath = args.output_dir / filename
-            if not destpath.exists():
-                srcpath = base.parent / 'reports' / filename
-                copyfile(srcpath, destpath)
-
 
     search_file_multithreaded(args, predictor)
 
@@ -295,47 +201,6 @@ def parse_arguments(args_list: List[str]) -> Tuple[argparse.Namespace,
     return known_args, unknown_args, parser
 
 
-def produce_index(args: argparse.Namespace, predictor: TacticPredictor,
-                  report_dir: Path,
-                  report_stats: List[search_report.ReportStats]) -> None:
-    predictorOptions = predictor.getOptions()
-    commit, date, weightshash = get_metadata(args)
-    search_report.write_summary(args, report_dir,
-                                predictorOptions +
-                                [("report type", "search"),
-                                 ("search width", args.search_width),
-                                 ("search depth", args.search_depth)],
-                                predictor.unparsed_args,
-                                commit, date, weightshash, report_stats)
-
-
-def stats_from_blocks(blocks: List[DocumentBlock], vfilename: str) \
-      -> search_report.ReportStats:
-    num_proofs = 0
-    num_proofs_failed = 0
-    num_proofs_completed = 0
-    for block in blocks:
-        if isinstance(block, ProofBlock):
-            if block.status != SearchStatus.SKIPPED:
-                num_proofs += 1
-            if block.status == SearchStatus.SUCCESS:
-                num_proofs_completed += 1
-            elif block.status == SearchStatus.FAILURE:
-                num_proofs_failed += 1
-    return search_report.ReportStats(vfilename, num_proofs,
-                                     num_proofs_failed, num_proofs_completed)
-
-
-def get_metadata(args: argparse.Namespace) -> Tuple[str, datetime.datetime, str]:
-    cur_commit = subprocess.check_output(["git show --oneline | head -n 1"],
-                                         shell=True).decode('utf-8').strip()
-    cur_date = datetime.datetime.now()
-    if args.weightsfile:
-        weights_hash = str(subprocess.check_output(
-            ["sha256sum", args.weightsfile]))
-    else:
-        weights_hash = ""
-    return cur_commit, cur_date, weights_hash
 
 
 def get_predictor(parser: argparse.ArgumentParser,
@@ -808,307 +673,7 @@ def search_file_multithreaded(args: argparse.Namespace,
         for worker in workers:
             worker.join()
     if args.generate_report:
-        generate_report(args, predictor)
-
-def generate_report(args: argparse.Namespace, predictor: TacticPredictor) -> None:
-    stats: List[search_report.ReportStats] = []
-    model_name = dict(predictor.getOptions())["predictor"]
-
-    project_dicts = project_dicts_from_args(args)
-    for project_dict in project_dicts:
-        for filename in project_dict["test_files"]:
-            file_solutions = []
-            output_file_prefix = args.output_dir / project_dict["project_name"] / \
-                  (util.safe_abbrev(Path(filename),
-                                    [Path(path) for path in
-                                     project_dict["test_files"]]))
-            source_file = args.prelude / project_dict["project_name"] / filename
-            try:
-                with (Path(str(output_file_prefix) + "-proofs.txt")).open('r') as f:
-                    for line in f:
-                        job, sol = json.loads(line)
-                        file_solutions.append((job, SearchResult.from_dict(sol)))
-            except FileNotFoundError:
-                cmds = serapi_instance.load_commands(source_file)
-                lemmas = serapi_instance.lemmas_in_file(source_file, cmds, args.include_proof_relevant)
-                assert len(lemmas) == 0
-                stats.append(search_report.ReportStats(filename, 0, 0, 0))
-                continue
-            blocks = blocks_from_scrape_and_sols(
-                source_file,
-                [(lemma_stmt, module_name, sol)
-                for (project, filename, module_name, lemma_stmt), sol
-                in file_solutions])
-
-            write_solution_vfile(args, output_file_prefix.with_suffix(".v"),
-                                 model_name, blocks)
-            write_html(args, output_file_prefix.with_suffix(".html"),
-                       filename, blocks)
-            write_csv(args, output_file_prefix.with_suffix(".csv"), blocks)
-            stats.append(stats_from_blocks(blocks, str(filename)))
-        produce_index(args, predictor,
-                      args.output_dir / project_dict["project_name"],
-                      stats)
-    if len(project_dicts) > 1:
-        multi_project_report.multi_project_index(args.output_dir)
-
-def blocks_from_scrape_and_sols(
-        src_filename: Path2,
-        lemma_statements_done: List[Tuple[str, str, SearchResult]]
-        ) -> List[DocumentBlock]:
-
-    interactions = data.read_all_text_data(
-        src_filename.with_suffix(".v.scrape"))
-
-    def lookup(module: str, lemma_stmt: str) -> Optional[SearchResult]:
-        for lstmt, lmod, lresult in lemma_statements_done:
-            if (lmod == module and
-                    serapi_instance.kill_comments(lstmt).strip()
-                    == serapi_instance.kill_comments(lemma_stmt).strip()):
-                return lresult
-        return None
-
-    def generate():
-        cur_lemma_stmt = ""
-        unique_lemma_stmt = ""
-
-        sm_stack = serapi_instance.initial_sm_stack(src_filename)
-
-        tactics_interactions_batch: List[TacticInteraction] = []
-        vernac_cmds_batch: List[str] = []
-
-        in_proof = False
-        obl_num = 0
-        last_program_statement = ""
-
-        def yield_proof():
-            nonlocal sm_stack
-            nonlocal tactics_interactions_batch
-            nonlocal cur_lemma_stmt
-            nonlocal unique_lemma_stmt
-            nonlocal in_proof
-            nonlocal obl_num
-            nonlocal last_program_statement
-
-            sm_prefix = serapi_instance.sm_prefix_from_stack(sm_stack)
-            batch_without_brackets = [t for t in tactics_interactions_batch
-                                      if t.tactic.strip() != "{" and
-                                      t.tactic.strip() != "}"]
-            result = lookup(sm_prefix, unique_lemma_stmt)
-            if result is None:
-                return ProofBlock(cur_lemma_stmt, sm_prefix,
-                                  SearchStatus.SKIPPED, [],
-                                  batch_without_brackets)
-            else:
-                return ProofBlock(cur_lemma_stmt, sm_prefix,
-                                  result.status, result.commands,
-                                  batch_without_brackets)
-            tactics_interactions_batch = []
-
-        for interaction in interactions:
-            if in_proof and isinstance(interaction, str):
-                in_proof = False
-                yield yield_proof()
-            elif in_proof and isinstance(interaction, ScrapedTactic):
-                tactics_interactions_batch.append(
-                    interaction_from_scraped(interaction))
-            elif isinstance(interaction, ScrapedTactic):
-                assert not in_proof
-                cur_lemma_stmt = vernac_cmds_batch[-1]
-                if re.match(r"\s*Next\s+Obligation\s*\.\s*",
-                            serapi_instance.kill_comments(
-                                cur_lemma_stmt).strip()):
-                    unique_lemma_stmt = \
-                      f"{last_program_statement} Obligation {obl_num}."
-                    obl_num += 1
-                else:
-                    unique_lemma_stmt = cur_lemma_stmt
-                yield VernacBlock(vernac_cmds_batch[:-1])
-                vernac_cmds_batch = []
-                tactics_interactions_batch = []
-                tactics_interactions_batch.append(
-                    interaction_from_scraped(interaction))
-                in_proof = True
-            if isinstance(interaction, str):
-                sm_stack = serapi_instance.update_sm_stack(sm_stack, interaction)
-                vernac_cmds_batch.append(interaction)
-                if re.match(r"\s*Program\s+.*",
-                            serapi_instance.kill_comments(interaction).strip()):
-                    last_program_statement = interaction
-                    obl_num = 0
-        if in_proof:
-            yield yield_proof()
-        pass
-    blocks = list(generate())
-    return blocks
-
-
-def interaction_from_scraped(s: ScrapedTactic) -> TacticInteraction:
-    return TacticInteraction(s.tactic, s.context)
-
-
-def write_solution_vfile(args: argparse.Namespace, output_filename: Path2,
-                         model_name: str,
-                         doc_blocks: List[DocumentBlock]):
-    with output_filename.open('w') as sfile:
-        for k, v in [("search-width", args.search_width),
-                     ("search-depth", args.search_depth),
-                     ("model", model_name)]:
-            print(f"(* {k}: {v} *)", file=sfile)
-
-        for block in doc_blocks:
-            if isinstance(block, VernacBlock):
-                for cmd in block.commands:
-                    print(cmd, file=sfile, end='')
-            else:
-                assert isinstance(block, ProofBlock)
-                print(block.lemma_statement, end='', file=sfile)
-                if block.predicted_tactics:
-                    for tac in block.predicted_tactics:
-                        print(tac.tactic, file=sfile)
-                else:
-                    for tac in block.original_tactics:
-                        print(tac.tactic, file=sfile)
-
-        pass
-
-
-def html_header(tag: Tag, doc: Doc, text: Text, css: List[str],
-                javascript: List[str], title: str) -> None:
-    with tag('head'):
-        for filename in css:
-            doc.stag('link', href=filename, rel='stylesheet')
-        for filename in javascript:
-            with tag('script', type='text/javascript',
-                     src=filename):
-                pass
-        with tag('title'):
-            text(title)
-
-
-def write_csv(args: argparse.Namespace,
-              output_filename: Path2,
-              doc_blocks: List[DocumentBlock]):
-    with output_filename.open('w', newline='') as csvfile:
-        for k, v in vars(args).items():
-            csvfile.write("# {}: {}\n".format(k, v))
-
-        rowwriter = csv.writer(csvfile, lineterminator=os.linesep)
-        for block in doc_blocks:
-            if isinstance(block, ProofBlock):
-                rowwriter.writerow([block.lemma_statement.strip(),
-                                    block.status,
-                                    len(block.original_tactics)])
-
-
-def write_html(args: argparse.Namespace,
-               output_file: Path2, filename: Path2,
-               doc_blocks: List[DocumentBlock]) -> None:
-    global unnamed_goal_number
-    unnamed_goal_number = 0
-    doc, tag, text, line = Doc().ttl()
-    with tag('html'):
-        html_header(tag, doc, text, [details_css], [details_javascript],
-                    "Proverbot Detailed Report for {}".format(str(filename)))
-        with tag('body', onload='init()'), tag('pre'):
-            for block_idx, block in enumerate(doc_blocks):
-                if isinstance(block, VernacBlock):
-                    write_commands(block.commands, tag, text, doc)
-                else:
-                    assert isinstance(block, ProofBlock)
-                    status_klass = classFromSearchStatus(block.status)
-                    write_lemma_button(block.lemma_statement, block.module,
-                                       status_klass, tag, text)
-                    with tag('div', klass='region'):
-                        with tag('div', klass='predicted'):
-                            write_tactics(args, block.predicted_tactics,
-                                          block_idx,
-                                          tag, text, doc)
-                        with tag('div', klass='original'):
-                            write_tactics(args, block.original_tactics,
-                                          block_idx,
-                                          tag, text, doc)
-    with output_file.open('w') as fout:
-        fout.write(doc.getvalue())
-
-
-def write_lemma_button(lemma_statement: str, module: Optional[str],
-                       status_klass: str, tag: Tag, text: Text):
-    global unnamed_goal_number
-    lemma_name = \
-        serapi_instance.lemma_name_from_statement(lemma_statement)
-    if module:
-        module_prefix = escape_lemma_name(module)
-    else:
-        module_prefix = ""
-    if lemma_name == "":
-        unnamed_goal_number += 1
-        fullname = module_prefix + lemma_name + str(unnamed_goal_number)
-    else:
-        fullname = module_prefix + lemma_name
-    if status_klass != "skipped":
-        with tag('button', klass='collapsible {}'.format(status_klass),
-                 onmouseover="hoverLemma(\"{}\")".format(fullname),
-                 onmouseout="unhoverLemma(\"{}\")".format(fullname)):
-            with tag('code', klass='buttontext'):
-                text(lemma_statement.strip())
-    else:
-        with tag('button', klass='collapsible {}'.format(status_klass)):
-            with tag('code', klass='buttontext'):
-                text(lemma_statement.strip())
-
-
-def write_commands(commands: List[str], tag: Tag, text: Text, doc: Doc):
-    for cmd in commands:
-        with tag('code', klass='plaincommand'):
-            text(cmd.strip("\n"))
-        doc.stag('br')
-
-
-def escape_quotes(term: str):
-    return re.sub("\"", "\\\"", term)
-
-
-def subgoal_to_string(args: argparse.Namespace, sg: Obligation) -> str:
-    return "(\"" + escape_quotes(sg.goal[:args.max_print_term]) + "\", (\"" + \
-        "\",\"".join([escape_quotes(hyp[:args.max_print_term]) for hyp in
-                      sg.hypotheses[:args.max_print_hyps]]) + "\"))"
-
-
-def write_tactics(args: argparse.Namespace,
-                  tactics: List[TacticInteraction],
-                  region_idx: int,
-                  tag: Tag, text: Text, doc: Doc):
-    for t_idx, t in enumerate(tactics):
-        idStr = '{}-{}'.format(region_idx, t_idx)
-        subgoals_str = "(" + ",".join([subgoal_to_string(args, subgoal)
-                                       for subgoal in
-                                       t.context_before.all_goals[
-                                           :args.max_print_subgoals]]) + ")"
-        with tag('span',
-                 ('data-subgoals', subgoals_str),
-                 id='command-{}'.format(idStr),
-                 onmouseover='hoverTactic("{}")'.format(idStr),
-                 onmouseout='unhoverTactic()'):
-            with tag('code', klass='plaincommand'):
-                text(t.tactic.strip())
-            doc.stag('br')
-
-
-def classFromSearchStatus(status: SearchStatus) -> str:
-    if status == SearchStatus.SUCCESS:
-        return 'good'
-    elif status == SearchStatus.INCOMPLETE:
-        return 'okay'
-    elif status == SearchStatus.SKIPPED:
-        return 'skipped'
-    elif status == SearchStatus.FAILURE:
-        return 'bad'
-    elif status == SearchStatus.CRASHED:
-        return 'bad'
-    else:
-        assert False
-
+        search_report.generate_report(args, predictor, project_dicts)
 
 def try_run_prelude(args: argparse.Namespace, coq: SerapiInstance):
     if not args.weightsfile:
